+++ Package.swift
@@ -7,7 +7,7 @@ let package = Package(
     name: "AgoraUIKit_iOS",
     platforms: [.iOS(.v13)],
     products: [
-        .library(name: "AgoraUIKit_iOS", targets: ["AgoraUIKit_iOS"]),
+        .library(name: "AgoraUIKit_iOS", targets: ["AgoraUIKit_iOS"])
     ],
     dependencies: [.package(
         name: "AgoraRtcKit",
