+++ Sources/Agora-UIKit/AgoraSingleVideoView.swift
@@ -12,7 +12,6 @@ import AppKit
 #endif
 import AgoraRtcKit
 
-
 /// View for the individual Agora Camera Feed.
 public class AgoraSingleVideoView: MPView {
     /// Is the video turned off for this user.
@@ -59,51 +58,6 @@ public class AgoraSingleVideoView: MPView {
         case microphone
     }
 
-    /// Find the string for the option ready to request the remote user to mute or unmute their mic or camera
-    /// - Parameters:
-    ///   - option: Device to be muted or umuted
-    ///   - isMuted: Boolean option to mute or unmute device
-    /// - Returns: String to be displayed in the mute/unmute option
-    open func userOptionsString(
-        for option: StreamMessageController.MutingDevices, isMuted: Bool
-    ) -> String {
-        switch option {
-        case .camera:
-            return isMuted ? MPButton.unmuteCameraString : MPButton.muteCameraString
-        case .microphone:
-            return isMuted ? MPButton.unmuteMicString : MPButton.muteMicString
-        }
-    }
-
-    func updateUserOptions() {
-        #if os(macOS)
-        guard let userOptions = self.userOptions as? NSPopUpButton else {
-            return
-        }
-        userOptions.removeAllItems()
-        self.addItems(to: userOptions)
-        #endif
-    }
-    #if os(macOS)
-    open func addItems(to userOptionsBtn: NSPopUpButton) {
-        let actionItem = NSMenuItem()
-        actionItem.attributedTitle = NSAttributedString(
-            string: "􀣋",
-            attributes: [ NSAttributedString.Key.foregroundColor: self.micFlagColor ]
-        )
-        userOptionsBtn.menu?.insertItem(actionItem, at: 0)
-        StreamMessageController.MutableDevices.allCases.forEach { enumCase in
-            var isMuted: Bool
-            switch enumCase {
-            case .camera:
-                isMuted = self.videoMuted
-            case .microphone:
-                isMuted = self.audioMuted
-            }
-            userOptionsBtn.addItem(withTitle: self.userOptionsString(for: enumCase, isMuted: isMuted))
-        }
-    }
-    #endif
     lazy var userOptions: MPView? = {
         #if os(iOS)
         let userOptionsBtn = MPButton.newToggleButton(
@@ -116,7 +70,7 @@ public class AgoraSingleVideoView: MPView {
 
 //        userOptionsBtn.wantsLayer = true
 //        userOptionsBtn.layer?.backgroundColor = .white
-        (userOptionsBtn.cell as! NSButtonCell).backgroundColor = .selectedContentBackgroundColor
+        (userOptionsBtn.cell as? NSButtonCell)?.backgroundColor = .selectedContentBackgroundColor
         self.addItems(to: userOptionsBtn)
         #endif
         self.addSubview(userOptionsBtn)
@@ -143,66 +97,10 @@ public class AgoraSingleVideoView: MPView {
         return userOptionsBtn
     }()
 
-    #if os(iOS)
-    /// The options button has been selected
-    /// - Parameter sender: Button that was selected
-    @objc open func optionsBtnSelected(sender: UIButton) {
-        let alert = UIAlertController(title: "Request Action", message: nil, preferredStyle: .actionSheet)
-        StreamMessageController.MutingDevices.allCases.forEach { enumCase in
-            var isMuted: Bool
-            switch enumCase {
-            case .camera:
-                isMuted = self.videoMuted
-            case .microphone:
-                isMuted = self.audioMuted
-            }
-            alert.addAction(UIAlertAction(title: self.userOptionsString(for: enumCase, isMuted: isMuted), style: .default, handler: optionsActionSelected(sender:)))
-        }
-        alert.addAction(UIAlertAction(title: "Cancel", style: .cancel, handler: nil))
-        self.streamContainer?.presentAlert(alert: alert, animated: true)
-    }
-
-    /// Action selected such as mute/unmute microphone/camera.
-    /// - Parameter sender: UIAlertAction that was selected.
-    open func optionsActionSelected(sender: UIAlertAction) {
-        guard let actionTitle = sender.title else { return }
-        if let reqError = self.streamContainer?.streamController?.createRequest(to: self.uid, fromString: actionTitle),
-           !reqError {
-            AgoraVideoViewer.agoraPrint(.error, message: "invalid action title: \(actionTitle)")
-        }
-    }
-    #else
-    /// Options button has been selected, now display available requests
-    /// - Parameter sender: Button that was selected
-    @objc public func optionsBtnSelected(sender: NSPopUpButton) {
-        guard let selectedType = UserOptions(rawValue: sender.selectedItem?.title ?? "") else {
-            return
-        }
-        switch selectedType {
-        case .camera:
-            self.streamContainer?.streamController?.sendMuteRequest(to: self.uid, mute: true, device: .camera)
-        case .microphone:
-            self.streamContainer?.streamController?.sendMuteRequest(to: self.uid, mute: true, device: .microphone)
-        default:
-            return
-        }
-//        let pop = actionbutt
-//
-//        pop.addItem(NSMenuItem(title: "Mute User", action: #selector(muteRequest), keyEquivalent: "m"))
-//        sender.addSubview(pop)
-    }
-    #endif
-
-    @objc func muteRequest() {
-    }
     /// Icon to show if this user is muting their microphone
     lazy var mutedFlag: MPView = {
         #if os(iOS)
-        let muteFlag = UIImageView(
-            image: UIImage(
-                systemName: MPButton.micSlashSymbol
-            )
-        )
+        let muteFlag = UIImageView(image: UIImage(systemName: MPButton.micSlashSymbol))
         muteFlag.tintColor = self.micFlagColor
         #else
         let muteFlag = MPButton()
@@ -238,7 +136,10 @@ public class AgoraSingleVideoView: MPView {
     ///   - micColor: Color to be applied when the local or remote user mutes their microphone
     ///   - showOptions: Whether we want to show options to mute/unmute this user
     ///   - streamContainer: Container to access the StreamMessageContainer.
-    public init(uid: UInt, micColor: MPColor, showOptions: Bool = false, streamContainer: StreamMessageContainer? = nil) {
+    public init(
+        uid: UInt, micColor: MPColor, showOptions: Bool = false,
+        streamContainer: StreamMessageContainer? = nil
+    ) {
         self.canvas = AgoraRtcVideoCanvas()
         self.micFlagColor = micColor
         self.streamContainer = streamContainer
